Move all population traits into a single trait
