+++ src/genetic.rs
@@ -1,4 +1,5 @@
 pub mod algorithms;
 pub mod genome;
+pub mod organism;
 pub mod population;
 pub mod traits;
